Generate an XML document of the contributions

The function generates an XML nodeset that contains the contributors' name, affiliation, and their CRediT roles with a structure outlined in the JATS 1.2 DTD specifications (eLife). The output is generated from an infosheet validated with the validate_infosheet function. The infosheet must be based on the infosheet_template.

print_xml(infosheet)

Arguments

infosheet
validated infosheet

Value

The function returns an xml nodeset containing the contributors listed for each CRediT role they partake in.

Warning

The function is primarily developed to be the part of a shiny app. As the validation is handled inside of the app separately, the function can break with non-informative errors if running locally without first validating it.

See also

Examples

validate_infosheet(infosheet = infosheet_template)

#> $missing_surname
#> $missing_surname$type
#> [1] "success"
#> 
#> $missing_surname$message
#> [1] "There are no missing surnames."
#> 
#> 
#> $missing_firstname
#> $missing_firstname$type
#> [1] "success"
#> 
#> $missing_firstname$message
#> [1] "There are no missing firstnames."
#> 
#> 
#> $duplicate_names
#> $duplicate_names$type
#> [1] "success"
#> 
#> $duplicate_names$message
#> [1] "There are no duplicate names in the infosheet."
#> 
#> 
#> $duplicate_initials
#> $duplicate_initials$type
#> [1] "success"
#> 
#> $duplicate_initials$message
#> [1] "There are no duplicate initials in the infosheet."
#> 
#> 
#> $missing_order
#> $missing_order$type
#> [1] "success"
#> 
#> $missing_order$message
#> [1] "There are no missing values in the order of publication."
#> 
#> 
#> $duplicate_order
#> $duplicate_order$type
#> [1] "success"
#> 
#> $duplicate_order$message
#> [1] "There are no duplicated order numbers in the infosheet."
#> 
#> 
#> $missing_affiliation
#> $missing_affiliation$type
#> [1] "success"
#> 
#> $missing_affiliation$message
#> [1] "There are no missing affiliations in the infosheet."
#> 
#> 
#> $missing_corresponding
#> $missing_corresponding$type
#> [1] "success"
#> 
#> $missing_corresponding$message
#> [1] "There is at least one author indicated as corresponding author."
#> 
#> 
#> $missing_credit
#> $missing_credit$type
#> [1] "success"
#> 
#> $missing_credit$message
#> [1] "All authors have at least one CRediT statement checked."
#> 
#> 
#> $missing_email
#> $missing_email$type
#> [1] "warning"
#> 
#> $missing_email$message
#> There is no email address provided for the corresponding author(s): 2
#> 
#> 

print_xml(infosheet = infosheet_template)

#> {xml_document}
#> <contrib-group>
#> [1] <contrib>\n  <name surname="Aczel" given-names="Balazs"/>\n  <role vocab= ...
#> [2] <contrib>\n  <name surname="Holcombe" given-names="Alex"/>\n  <role vocab ...
#> [3] <contrib>\n  <name surname="Kovacs" given-names="Marton"/>\n  <role vocab ...